working on configuration testing
