+++ clients/python-client/README.md
@@ -96,3 +96,12 @@ chunkr = Chunkr(
 )
 ```
 
+## Run tests
+
+```python
+# Install dependencies
+uv pip install -e ".[test]"
+
+# Run tests
+uv run pytest
+```
