+++ clients/python-client/tests/test_chunkr.py
@@ -1,10 +1,10 @@
 import pytest
 from pathlib import Path
 from PIL import Image
-from chunkr_ai import Chunkr, ChunkrAsync
-from chunkr_ai.models import TaskResponse
 
-# Test fixtures
+from chunkr_ai import Chunkr, ChunkrAsync
+from chunkr_ai.models import Configuration, OcrStrategy, TaskResponse
+
 @pytest.fixture
 def chunkr():
     return Chunkr()
@@ -15,11 +15,11 @@ def async_chunkr():
 
 @pytest.fixture
 def sample_path():
-    return Path("tests/test.pdf")
+    return Path("tests/files/test.pdf")
     
 @pytest.fixture
 def sample_image():
-    img = Image.new('RGB', (100, 100), color='red')
+    img = Image.open("tests/files/test.jpg")
     return img
 
 def test_send_file_path(chunkr, sample_path):
@@ -28,6 +28,7 @@ def test_send_file_path(chunkr, sample_path):
     assert isinstance(response, TaskResponse)
     assert response.task_id is not None
     assert response.status == "Succeeded"
+    assert response.output is not None
 
 def test_send_file_path_str(chunkr, sample_path):
     response = chunkr.upload(str(sample_path))
@@ -35,6 +36,7 @@ def test_send_file_path_str(chunkr, sample_path):
     assert isinstance(response, TaskResponse)
     assert response.task_id is not None
     assert response.status == "Succeeded"
+    assert response.output is not None
 
 def test_send_opened_file(chunkr, sample_path):
     with open(sample_path, 'rb') as f:
@@ -43,10 +45,30 @@ def test_send_opened_file(chunkr, sample_path):
     assert isinstance(response, TaskResponse)
     assert response.task_id is not None
     assert response.status == "Succeeded"
-
-def test_send_pil_image(chunkr, sample_image):
-    response = chunkr.upload(sample_image)
+    assert response.output is not None
     
+# def test_send_pil_image(chunkr, sample_image):
+#     response = chunkr.upload(sample_image)
+    
+#     assert isinstance(response, TaskResponse)
+#     assert response.task_id is not None
+#     assert response.status == "Succeeded"
+
+def test_ocr_auto(chunkr, sample_path):
+    response = chunkr.upload(sample_path, Configuration(
+        ocr_strategy=OcrStrategy.AUTO
+    ))
     assert isinstance(response, TaskResponse)
     assert response.task_id is not None
     assert response.status == "Succeeded"
+    assert response.output is not None
+
+def test_ocr_expires_in(chunkr, sample_path):
+    response = chunkr.upload(sample_path, Configuration(
+        expires_in=10
+    ))
+    assert isinstance(response, TaskResponse)
+    assert response.task_id is not None
+    assert response.status == "Succeeded"
+    assert response.output is not None
+
