+++ core/src/routes/task.rs
@@ -45,7 +45,11 @@ pub async fn get_task_status(
         Ok(task_response) => Ok(HttpResponse::Ok().json(task_response)),
         Err(e) => {
             eprintln!("Error getting task status: {:?}", e);
-            Ok(HttpResponse::InternalServerError().body(e.to_string()))
+            if e.to_string().contains("has expired") || e.to_string().contains("not found") {
+                Ok(HttpResponse::NotFound().body("Task not found"))
+            } else {
+                Ok(HttpResponse::InternalServerError().body(e.to_string()))
+            }
         }
     }
 }
